# docs/swagger.yaml
definitions:
  controller.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  model.ListParams:
    properties:
      page:
        type: integer
      size:
        type: integer
    type: object
  model.PostDTOAdd:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  model.PostDTOInsert:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  model.PostDTOUpdateByAdmin:
    properties:
      content:
        type: string
      post_id:
        example: "0"
        type: string
      title:
        type: string
    required:
    - content
    - post_id
    - title
    type: object
  model.PostDTOUpdateBySelf:
    properties:
      content:
        type: string
      post_id:
        example: "0"
        type: string
      title:
        type: string
      user_id:
        example: "0"
        type: string
    required:
    - content
    - post_id
    - title
    - user_id
    type: object
...
info:
  contact: {}
paths:
  /post/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新增的文章参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.PostDTOAdd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 通过获取文章视图列表
      tags:
      - 文章相关接口
  /post/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要删除的post_id
        in: query
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 通过post_id删除文章
      tags:
      - 文章相关接口
  /post/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要查询的post_id
        in: query
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员通过post_id获取文章全部数据
      tags:
      - 文章相关接口
  /post/get/vo:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要查找的文章id
        in: query
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 通过post_id获取文章视图
      tags:
      - 文章相关接口
  /post/list/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分页查询所需要的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员获取全部文章详细信息
      tags:
      - 文章相关接口
  /post/list/page/vo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分页查询需要的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 通过获取文章视图列表
      tags:
      - 文章相关接口
  /post/my:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分页查询需要的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 通过当前登录的用户所写的文章
      tags:
      - 文章相关接口
  /post/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新建文章参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.PostDTOInsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 新建文章
      tags:
      - 文章相关接口
  /post/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要更新的文章信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.PostDTOUpdateByAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员编辑文章
      tags:
      - 文章相关接口
  /post/update/my:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      - description: 修改后的数据
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.PostDTOUpdateBySelf'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户更新自己写的文章
      tags:
      - 文章相关接口
...
swagger: "2.0"